Move CreatePullRequest to own component
